🐛 fix: Allow postgres to properly query timestamps
🧾 Testing with postgre database, I ran into some problems where the
timestamp was not formated corretly.  It turns out that I was using the
Datetime for the column when I should have been using Time. I've updated
"Initialze tables" script to include this new format.  I've also set the
settings varialbe to be postgresql instead of postgres.  This should
be part of the ettings