Package (#2)
* Making a package changing dir and fixing docker

* Adding Logging conf back in

* Messing around with base/data dir

* Updating models and packages

* Updating some calls to logger and settings. Updating notifier options

* Create docker-compose.dev.yml

* Update Taskfile.yml

* Updating BASE_DIR

* Updating/Testing email, forgot password etc

* Updating some password reset, tokens etc

* ✨ feat        – new feature
📣 Template directory f/ custom template. Plus email fixes

🧾 Adding the ability to add a branded or custom template to the
data/templates directory. So you dont have to add it to the source code.
You can mount a directory or make custom changes without effecting the
code.

* 🐛 fix: Allow postgres to properly query timestamps

🧾 Testing with postgre database, I ran into some problems where the
timestamp was not formated corretly.  It turns out that I was using the
Datetime for the column when I should have been using Time. I've updated
"Initialze tables" script to include this new format.  I've also set the
settings varialbe to be postgresql instead of postgres.  This should
be part of the ettings

* 📚 refactor: Use the feature_details in settings

🧾 Utilize the feature_detailts function in settings to disable or enable
the plugins feature.  Also added SecretStr for secrets in pydantic models